# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Leedz API
  description: |
    API for Leedz - Sales Intelligence & CRM Platform
    
    ## Authentication
    All endpoints (except public ones) require a Bearer token in the Authorization header.
    
    ## Tenant Context
    The current tenant is determined from the JWT token. Use `/api/auth/switch-tenant` to change tenants.
    
    ## Rate Limiting
    - Default: 100 requests/minute
    - Search: 10 requests/minute
    - WhatsApp: 100 requests/hour
  version: 2.0.0
  contact:
    name: Leedz Team
    email: dev@leadz.sa
servers:
  - url: https://api.leadz.sa
    description: Production
  - url: https://api.staging.leadz.sa
    description: Staging
  - url: http://localhost:3000
    description: Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: object
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatarUrl:
          type: string
          nullable: true
        emailVerified:
          type: boolean
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
    UserRole:
      type: string
      enum: [OWNER, ADMIN, MANAGER, SALES]
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            tenant:
              $ref: '#/components/schemas/Tenant'
            token:
              type: string
            expiresIn:
              type: integer
    SignupRequest:
      type: object
      required:
        - organizationName
        - name
        - email
        - password
      properties:
        organizationName:
          type: string
          minLength: 2
          maxLength: 100
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        phone:
          type: string
          nullable: true

    # ==================== Invite ====================
    Invite:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          type: string
          enum: [PENDING, ACCEPTED, EXPIRED, CANCELLED]
        message:
          type: string
          nullable: true
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    # ==================== Leads ====================
...
    EvidenceItem:
      type: object
      required:
        - type
        - title
        - source
        - snippet
        - confidence
        - hash
        - collectedAt
        - sizeBytes
      properties:
        type:
          type: string
          enum:
            - GOOGLE_MAPS_LISTING
            - WEBSITE_CONTENT
            - SOCIAL_PROFILE
            - SEARCH_RESULT
            - CONTACT_INFO
            - REVIEW
            - NEWS_ARTICLE
        title:
          type: string
          maxLength: 255
        source:
          type: string
          description: Connector that collected this evidence
        url:
          type: string
          format: uri
        snippet:
          type: string
          maxLength: 10000
          description: Text-only content, sanitized
        confidence:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        rawData:
          type: object
          description: Structured data if available, max 50KB
        hash:
          type: string
          description: SHA-256 of snippet for deduplication
        collectedAt:
          type: string
          format: date-time
        sizeBytes:
          type: integer
          description: For quota tracking
        pageLoadTimeMs:
          type: integer
          description: For debugging

    # ==================== Audit ====================
...
paths:
...
  /api/auth/signup:
    post:
      tags: [Auth]
      summary: Create new organization and owner account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login to existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user and tenant info
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      tenant:
                        $ref: '#/components/schemas/Tenant'
                      tenants:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tenant'
...
  /api/invites:
    get:
      tags: [Team]
      summary: List pending invites
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of invites
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      invites:
                        type: array
                        items:
                          $ref: '#/components/schemas/Invite'
    post:
      tags: [Team]
      summary: Create invite
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - role
              properties:
                email:
                  type: string
                  format: email
                role:
                  $ref: '#/components/schemas/UserRole'
                message:
                  type: string
      responses:
        '201':
          description: Invite created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      invite:
                        $ref: '#/components/schemas/Invite'
  /api/invites/{token}/accept:
    post:
      tags: [Team]
      summary: Accept invite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - password
              properties:
                name:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        '201':
          description: Invite accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      tenant:
                        $ref: '#/components/schemas/Tenant'
                      token:
                        type: string
  /api/invites/{id}:
    delete:
      tags: [Team]
      summary: Cancel invite
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Invite cancelled
  /api/invites/{id}/resend:
    post:
      tags: [Team]
      summary: Resend invite
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Invite resent

  # ==================== Auth Switch Tenant (CF-11 fix) ====================
...
  /api/agent/config:
    get:
      tags: [Agent]
      summary: Get agent configuration (allowlist, feature flags, connector settings)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Agent configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      allowlist:
                        type: array
                        items:
                          type: string
                        description: Allowed domains for execution
                      featureFlags:
                        type: object
                        additionalProperties:
                          type: boolean
                      connectorSettings:
                        type: object
                        additionalProperties:
                          type: object
  /api/agent/heartbeat:
    post:
      tags: [Agent]
      summary: Agent heartbeat to maintain connection
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                extensionVersion:
                  type: string
                activeJobs:
                  type: array
                  items:
                    type: string
                status:
                  type: string
                  enum: [IDLE, BUSY, PAUSED]
      responses:
        '200':
          description: Heartbeat acknowledged
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      pendingJobs:
                        type: integer
                      serverTime:
                        type: string
                        format: date-time
  /api/agent/jobs/{jobId}/ack:
    post:
      tags: [Agent]
      summary: Acknowledge job receipt
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accepted:
                  type: boolean
                reason:
                  type: string
      responses:
        '200':
          description: Acknowledgment received
  /api/agent/jobs/{jobId}/progress:
    post:
      tags: [Agent]
      summary: Report job progress
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stepId:
                  type: integer
                progress:
                  type: integer
                  minimum: 0
                  maximum: 100
                message:
                  type: string
      responses:
        '200':
          description: Progress recorded
  /api/agent/jobs/{jobId}/evidence:
    post:
      tags: [Agent]
      summary: Submit evidence batch
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stepId:
                  type: integer
                evidence:
                  type: array
                  items:
                    $ref: '#/components/schemas/EvidenceItem'
      responses:
        '200':
          description: Evidence stored
        '413':
          description: Evidence too large
  /api/agent/jobs/{jobId}/error:
    post:
      tags: [Agent]
      summary: Report job error
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stepId:
                  type: integer
                code:
                  type: string
                  enum:
                    - LOGIN_REQUIRED_UNSUPPORTED
                    - RATE_LIMITED
                    - BLOCKED
                    - CAPTCHA_REQUIRED
                    - NOT_FOUND
                    - PROFILE_PRIVATE
                    - TIMEOUT
                    - NETWORK_ERROR
                    - PARSE_ERROR
                    - CONSENT_REQUIRED
                    - QUOTA_EXCEEDED
                message:
                  type: string
                recoverable:
                  type: boolean
      responses:
        '200':
          description: Error recorded
  /api/agent/jobs/{jobId}/done:
    post:
      tags: [Agent]
      summary: Mark job as complete
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [SUCCESS, FAILED, PARTIAL_SUCCESS, CANCELLED]
                summary:
                  type: object
                  properties:
                    stepsCompleted:
                      type: integer
                    stepsTotal:
                      type: integer
                    evidenceCount:
                      type: integer
                    duration:
                      type: integer
                      description: Duration in milliseconds
                error:
                  type: object
                  properties:
                    code:
                      type: string
                    message:
                      type: string
                    stepId:
                      type: integer
      responses:
        '200':
          description: Job marked complete

  # ==================== Audit ====================
...